+++ package/com.unity.formats.usd/Runtime/InitUsd.cs
@@ -28,7 +28,7 @@ namespace Unity.Formats.USD {
       m_usdInitialized = true;
 
       try {
-        // Initializes native USD plugins and ensures plugins are discoverable on the system path. 
+        // Initializes native USD plugins and ensures plugins are discoverable on the system path.
         SetupUsdPath();
 
         // Type registration enables automatic conversion from Unity-native types to USD types (e.g.
@@ -51,6 +51,8 @@ namespace Unity.Formats.USD {
     private static void SetupUsdPath()
     {
 #if UNITY_EDITOR
+      // TODO: this is not robust, e.g. if anyone changes CWD from the default, package resolution
+      // will fail. Following up with UPM devs to see what we can do about it.
       var supPath = System.IO.Path.GetFullPath("Packages/com.unity.formats.usd/Runtime/Plugins");
 #else
       var supPath = UnityEngine.Application.dataPath.Replace("\\", "/") + "/Plugins";
@@ -60,12 +62,16 @@ namespace Unity.Formats.USD {
       supPath += @"/x86_64/share/";
 #elif (UNITY_EDITOR_OSX)
       supPath += @"/x86_64/UsdCs.bundle/Contents/Resources/share/";
+#elif (UNITY_EDITOR_LINUX)
+      supPath += @"/x86_64/share";
 #elif (UNITY_STANDALONE_WIN)
       supPath += @"/share/";
 #elif (UNITY_STANDALONE_OSX)
       supPath += @"/UsdCs.bundle/Contents/Resources/share/";
+#elif (UNITY_STANDALONE_LINUX)
+      supPath += "/share";
 #endif
-      
+
       Debug.LogFormat("Registering plugins: {0}", supPath);
       pxr.PlugRegistry.GetInstance().RegisterPlugins(supPath);
     }