+++ package/com.unity.formats.usd/Samples/ImportProcessor/SetHideFlags.cs
@@ -24,21 +24,38 @@ namespace Unity.Formats.USD.Examples {
 
     public HideFlags hideFlagsSettings = HideFlags.DontSave;
 
+    public enum Operation {Overwrite, LogicalAND, LogicalOR, LogicalXOR}
+    [Tooltip("The logical operator to use when setting the flags. In most cases Or will produce expected results")]
+    public Operation operation = Operation.LogicalOR;
+
     public void PostProcessHierarchy(PrimMap primMap, SceneImportOptions sceneImportOptions) {
       InitRegex();
 
       foreach (KeyValuePair<SdfPath, GameObject> kvp in primMap) {
         if (!IsMatch(kvp.Key)) continue;
         GameObject go = kvp.Value;
-        go.hideFlags = hideFlagsSettings;
+        switch (operation) {
+          case Operation.LogicalAND:
+            go.hideFlags &= hideFlagsSettings;
+            continue;
+          case Operation.LogicalOR:
+            go.hideFlags |= hideFlagsSettings;
+            continue;
+          case Operation.LogicalXOR:
+            go.hideFlags ^= hideFlagsSettings;
+            continue;
+          default:
+            go.hideFlags = hideFlagsSettings;
+            continue;
+        }
       }
     }
 
     void Reset() {
-      matchExpression = "/";
-      isNot = true;
-      matchType = EMatchType.Wildcard;
-      compareAgainst = ECompareAgainst.UsdName;
+      matchExpression = @"^\/.+";
+      isNot = false;
+      matchType = EMatchType.Regex;
+      compareAgainst = ECompareAgainst.UsdPath;
     }
   }
 }