+++ src/USD.NET.Unity/Properties/AssemblyInfo.cs
@@ -8,9 +8,9 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTitle("USD.NET.Unity")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Google")]
+[assembly: AssemblyCompany("Unity Technologies")]
 [assembly: AssemblyProduct("USD.NET.Unity")]
-[assembly: AssemblyCopyright("Copyright ©2017 Google Inc")]
+[assembly: AssemblyCopyright("Copyright ©2019 Unity Technologies")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("19.01.0.*")]
-[assembly: AssemblyFileVersion("19.01.0.0")]
+[assembly: AssemblyVersion("19.05.0.*")]
+[assembly: AssemblyFileVersion("19.05.0.0")]