+++ src/USD.NET.Unity/Shading/UsdPreviewSurface/PrimvarReaderSample.cs
@@ -21,7 +21,7 @@ namespace USD.NET.Unity {
   /// </summary>
   [System.Serializable]
   [UsdSchema("Shader")]
-  public class PrimvarReaderSample<T> : ShaderSample {
+  public class PrimvarReaderSample<T> : ShaderSample where T : struct {
 
     public PrimvarReaderSample() {
       if (typeof(T) == typeof(float)) {
@@ -66,18 +66,13 @@ namespace USD.NET.Unity {
     [InputParameter("_Fallback")]
     public Connectable<T> fallback = new Connectable<T>();
 
-    // TODO(jcowles): outputs in the UsdPreviewSurface cannot have values assigned to them, however
-    //                USD .NET has no way to stop this from happening for value types currently. 
-    //                https://github.com/googlevr/usd-unity-sdk/issues/12
-
-    /*
     public class Outputs : SampleBase {
-      public T result;
+      public T? result;
     }
 
     [UsdNamespace("outputs")]
     public Outputs outputs = new Outputs();
-    */
+
   }
 
 }