+++ package/com.unity.formats.usd/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changes in usd-unitysdk for Unity
 
+## [1.0.1-preview.1] - 2019-06-30
+### Changes
+
+New feature :
+- Linux support
+
+Bug fixes :
+- NullReferenceException on USD export (#67)
+- Path of .usd file does not match textures in the archive (#88)
+- Build error when sample is imported
+- Timeline: Playback sometimes spews errors (#93)
+- Skeleton bindings below the root are not discovered (#92)
+- Normals update after opening standard shader (#91)
+- UsdSkel export fails when rig root is the model root (#89)
+- VariantSets on the root UsdAsset results in errors (#77)
+- Bones do not match bindpose (#98)
+- Standard shader "pops" when opening the inspector (#102)
+- UsdIo: Do not recurse on null objects (#118)
+
 ## [1.0.0-preview.4] - 2019-03-15
 ### Changes
 - Update readme images